Feature/mailer

+++ app/services/resources/resource_service.rb
@@ -1,7 +1,7 @@
 module Resources
   class ResourceService
     attr_reader :params, :user
-    attr_accessor :resource
+    attr_accessor :resource, :resources
 
     def initialize(params = {})
       @params = params
@@ -21,53 +21,94 @@ module Resources
       update_resource
     end
 
+    def execute_get_resources
+      get_resources
+    end
+
     private
 
     def create_resource
-      @resource = Resource.new(resource_params.merge(user_id: current_user.id))
-      if @resource.save!
-        success_response(@resource)
+      if current_user.present?
+        @resource = Resource.new(resource_params.merge(user_id: current_user.id, tenant_id: current_user.tenant_id))
+        if @resource.save!
+          success_response(@resource)
+        else
+          error_response(@resource.errors.full_messages)
+        end
       else
-        error_response(@resource.errors.full_messages)
+        error_response([ "User not authenticated" ])
       end
-    rescue ActiveRecord::RecordInvalid => e
+    rescue ActiveRecord::RecordNotFound => e
+      error_response([ e.message ])
+    rescue ActiveRecord::RecordNotCreated => e
+      error_response([ e.message ])
+    rescue StandardError => e
       error_response([ e.message ])
     end
 
     def delete_resource
       @resource = find_resource
-      # authorize_admin!
-      if @resource.destroy
-        success_response(@resource)
+      if current_user.present?
+        if @resource.destroy
+          success_response(@resource)
+        else
+          error_response(@resource.errors.full_messages)
+        end
       else
-        error_response(resource.errors.full_messages)
+        error_response([ "User not authenticated" ])
       end
+    rescue ActiveRecord::RecordNotFound => e
+      error_response([ e.message ])
+    rescue ActiveRecord::RecordNotDestroyed => e
+      error_response([ e.message ])
+    rescue StandardError => e
+      error_response([ e.message ])
     end
+
+
 
     def update_resource
       @resource = find_resource
-      # authorize_admin!
-      if @resource.update(resource_params)
-        success_response(@resource)
+      if current_user.present?
+        if @resource.update(resource_params)
+          success_response(@resource)
+        else
+          error_response(resource.errors.full_messages)
+        end
       else
-        error_response(resource.errors.full_messages)
+        error_response([ "User not authenticated" ])
+      end
+    rescue ActiveRecord::RecordNotFound => e
+      error_response([ e.message ])
+    rescue ActiveRecord::RecordNotDestroyed => e
+      error_response([ e.message ])
+    rescue StandardError => e
+      error_response([ e.message ])
+    end
+
+    def get_resources
+      begin
+        @resources = Resource.where(tenant_id: current_user.tenant_id).order(created_at: :DESC)
+        success_response(@resources)
+      rescue ActiveRecord::RecordNotFound => e
+        error_response([ e.message ])
+      rescue StandardError => e
+        error_response([ e.message ])
       end
     end
+
 
     def current_user
       current_user ||= params[:current_user]
     end
 
-    # def authorize_admin!
-    # raise StandardError.new("Unauthorized") unless user.admin?
-    # end
 
     def find_resource
       Resource.find_by!(id: params[:id])
     end
 
     def resource_params
-      ActionController::Parameters.new(params).permit(:resource_category, :resource_status, :tenant_id)
+      ActionController::Parameters.new(params).permit(:resource_category, :resource_status)
     end
 
     def success_response(resource)
